requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cities==0.6.1
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.3.1
django-reset-migrations==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.14
zipp==3.11.0
